+++ apps/tempest.games/src/frontend/<svg>.tsx
@@ -1,11 +1,14 @@
 import * as React from "react"
 
-export function tempest(): React.ReactNode {
+export function tempest(
+	props: React.SVGAttributes<SVGElement>,
+): React.ReactNode {
 	return (
 		<svg
 			id="uuid-3d2d6be6-f179-42a3-9511-1e9ddeb8575f"
 			xmlns="http://www.w3.org/2000/svg"
 			viewBox="0 0 500 101"
+			{...props}
 		>
 			<title>Tempest Games Logo</title>
 			<path
