+++ apps/tempest.games/src/library/alpha-rand.ts
@@ -1,0 +1,5 @@
+import { customAlphabet } from "nanoid"
+
+const alphabet = `abcdefghijklmnopqrstuvwxyz`
+
+export const alphaRand = customAlphabet(alphabet, 6)
